added cancelAuction method
this method will take an auction id and search for it in our auctions array. I'm not sure how we aim to implement this since we can either do it through ID or the auction directly, but for now I have it doing it by ID. Once it finds the auction using the ID, it will check if the auction has no bid, and if it doesn't have a bid, it will remove it. This method is currently commented out.